replace internal sound chip clocks with double and do proper rounding for sample rate part II
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5136 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sndintrf.h
@@ -450,7 +450,7 @@ const char *sound_name(const struct MachineSound *msound);
 /* returns number of chips, or 0 if the sound type doesn't support multiple instances */
 int sound_num(const struct MachineSound *msound);
 /* returns clock rate, or 0 if the sound type doesn't support a clock frequency */
-int sound_clock(const struct MachineSound *msound);
+double sound_clock(const struct MachineSound *msound);
 
 int sound_scalebufferpos(int value);
